requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
Markdown==3.3.4
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
openpyxl==3.1.2
pandas==2.0.2
pandas-stubs==2.0.2.230605
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tqdm==4.65.0
types-pytz==2023.3.0.0
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
